run_test now retuns a map of test nam->map name
What 0run reads the output of a test program one line at a time (we assume one
line == one test) and matches it against a list of named regular expressions.
The "match" (if any) of a particular test was the first regex in he list to
match.  In addition, the regex must pull out a string called the "name" of that
test.  In this we can match test results into groups like "passed" and
"failed".  The result is a dictionary mapping the test names to whatever
matched it.