run_main and it's Results.
run_main the high level test runner.  It reads argv and catches exceptions,
turning them into fatal, but readable.  It both scans the source and runs the
tests and then returns the results from both operations, in the form of a
bundle called a Result object.

Result objects hold the scan results, and also containsa number, errno which is
initially set to zero. Most importantly, it contains methods for checking the
validity of the results.  These methods are called as

        results.check_xyz(tset)

If all the members of "tset" are indeed in "xyz", all is fine, otherwise an
error message is printed to the terminal.  And results.errno is set to
something nonzero.

The check_ functions are deliberately designed as (text-based) UI helpers.  If
you want more flexibility, examine the data inside the results yourself.