Introducing the Runner class (and Fail_Runner)
The Runner has methods (any one of which the user might replace)
        * run,
        * scan_source and
        * scan_output.

The runner owns Popen object connecting us to the test process, which can be
examined by the run method.  One thing is that .run() must call wait on the
Popen, or else you get a zombie (not that it matters much).

The flexibility is used by Fail_Runner which actually runs the command, but
*wants* it to fail.  We can now run elm-fail  without seeing red-text.  Some
white text does leak through to stderr though.