Introducing RunData.
This class takes a source file and test command as constructor arguments and
just runs the command, storing stdout, stderr and "errno".  I had intended to
also store the text of the source, but forgot and ended up just storing the
filename.  This class exists because I think it is better to have a separate
type for descring half processed data than trying to track the state of a
mutable object as is processed.

The Runner class now creates a RunData object at initialisation.  As a result
the ".run()" method became misnamed.  So now we have split it's work in two
public, overridable functions "check_output" and "scan_output".

This has beeen used to clean

RawData class is now used and is called RunData.

RunData related cruft removal and renames.

The "Runner" now runs the command right at the start.  That means we have
ditche the ".run" method, but mad scan_output public.