New CLI for n0run.
Now it is
        n0run  source.c  ./test_command ...args...

So now the source file is mandator.  In return, you can pass arguments!
The change to elm-fail-run is a no-op, it just shows that we everything
still works if the run commands are lists instead of strings.

In fact we will now require that they be lists.