BUGFIX: proper callingof sub-0unit tests.
Sometimes it is convenient to wriet subroutines to checks which return zero or
one.   I have a habit of calling  them.

        chk_some_aspect()

This is all fine, but you must call it as

        CHK( chk_some_aspect() )

otherwise the return value will be ignored and the larger test will pass even
though the smaller one will fail.  This is probably not the end of the world,
you will still get red text, and n0run will still pick it up, but it is not
quite the idea.