Turned compiler errors back on again.
We had the makefile allow implicit function declations, because
some of them would be missing during the moving process.  Now
that the move is complete we turn this back on.

Also the whole point of splitting the tests out of elm.c was to make
sure 'elm.h' is actually complete.  It turns out that we were missing
'destroy_logger'.  Also, during the move process we added 'panic_nomem'
which had to go into the header.