Version numbering and version IDs.
We have defined the a format for version ID strings and provide the
compiletime ELM_VERSION and run-time elm_version() which can query the
current elm version.  In any given build these two things will be the
same string, but dynamic linking could change that.

We even have a unit test to sanity check that the version string is
valid.