Make install
Copies the library and headers to a unix-shaped tree under the
environment variable INSTALL_DIR.  If this is not defined, BUILD_DIR is
used.  I think elm0 should be part of the source tree of applications.

If those apps put the elm0 source in elm0/, they can do things
like:

        BUILD_DIR=../relative/build/path -C elm0/ install