The elm0 repository is now a submodule of itself.
Our problem: make the main source tree visible to the web site.

Github puts the web site in the same repository as the source, but in a
seprate branch.  Worse, both the website and the source contend for the
root directory.

Our solution: Pull elm0 in a submodule of itself.

Gitbhub says that if your pages tree has submodules, that they will be
pulled in and usable.  So by pulling elm0, we get to see whatever is on
the master branch.  Hopefully there is no infinite regress, the master
branch does not have any submodules.

Eeek.