Loggers have refcounts (static ones can be reffed)
Loggers now maintian have a refcount object.  For loggers created with
new_logger this is a true refcount which can be incremented with `ref_logger`.
`destroy_logger` will decrement the refcount and only destroy the logger on the
last ref.

For the four static loggers, `destroy_logger` and `ref_logger` are no-ops which
can safely be called any number of times.