error default: fwrite has default.
`error_fwrite` has changed its usage and interface.  It used to be a built-in
implementation for the .fwrite method.  It is now the wrapper that will either
call that method, or else supply a default behaviour.

This means you must not call the method directly - it might be NULL.
`error_fwrite` is called internally by `log_error`, but it is also and external
interface, in case you have your own reasons for writing errors to streams.