error default: type set in elm_mkerr, not init.
The idea is that the error type is set before the constructor is called, and
the constructor leaves it alone.  This way the same constructor can be used by
multiple types.

Along with this change, we make `elm_mkerr` set all unused fields to 0.  This
reflects a change of emphasis: previously the init_* function was the real
initialisar and `elm_mkerr` was only used instead of malloc() so that we could
sneak in the LINE, FUNCTION and FILE.  Now elm_mkerr is more of a true (if very
low level) initialiser that sets up defaults that the init_* can customise.