error default: ERROR_WITH and ERROR_ALLOC
The meaning of `ERROR_WITH` has now changed.  In the past it used to create an
error of arbitrary type, and would call the correct constructor - assuming you
followed the correct naming convention for constructors.  Now `ERROR_WITH`
always calls `init_error` and therefore should only be used with "simple
custom" errors, which are really the same as a standard error under the hood.

True custom errors should still have their own constuctors, but now the name
is just a convention.  Creating an error is done with

                init_your_error(ERROR_ALLOC(your_error), ....)

Of course, when you define a new error type you should wrap the above up in a
macro.