BUGFIX: ZALLOC is gone.
ZALLOC was our wrapper for the standard "calloc".  I call it a "bug" because
the most common use is stupid.  The temptation is to use it to initialise
structs to zero-default.  But this doesn't work on architectures where NULL is
not bytewise equal to the integer zero.  Arcane, and pedantic it's true, but
that explanation is more virtuous than saying "I can't be bothered testing it".